A lot of new things

+++ post_service/src/main/java/com/veggietalk/post_service/service/impl/PostServiceImpl.java
@@ -8,6 +8,7 @@ import com.veggietalk.post_service.service.converters.PostConverters;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.List;
@@ -17,6 +18,7 @@ import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
+@Transactional
 public class PostServiceImpl implements PostService {
 
     private final PostRepo postRepo;
@@ -25,7 +27,6 @@ public class PostServiceImpl implements PostService {
 
     DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd-MM-yyyy");
 
-    @Override
     public Post createPost(Post post) throws IllegalArgumentException{
         if (post.getDescription() == null || post.getDescription().isEmpty()){
             throw new IllegalArgumentException("There must be a description given for the post!");
@@ -34,18 +35,21 @@ public class PostServiceImpl implements PostService {
         return PostConverters.PostEntityConverter(postRepo.save(PostConverters.PostConverter(post)));
     }
 
-    @Override
     public void deletePost(Long id, Long userId) throws IllegalArgumentException{
         Optional<PostEntity> post = postRepo.findById(id);
         if(post.isPresent()){
-            postRepo.deletePost(id);
+            if(Objects.equals(post.get().getUser_id(), userId)){
+                postRepo.deletePost(id);
+            }
+            else{
+                throw new IllegalArgumentException("The account you are logged in cannot delete the selected post");
+            }
         }
         else{
             throw new IllegalArgumentException("The selected post does not exist :(");
         }
     }
 
-    @Override
     public List<Post> getAllPosts() throws NoSuchElementException {
         List<PostEntity> posts = postRepo.getAllPosts();
         if(!posts.isEmpty()){
